Remove use of TemplateHaskell extension
