+++ src/Data/Bifunctor/TH.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE CPP #-}
-{-# LANGUAGE TemplateHaskell #-}
 
 {-|
 Module:      Data.Bifunctor.TH
@@ -35,33 +34,18 @@ module Data.Bifunctor.TH (
     , makeBisequence
     ) where
 
-import           Control.Applicative
 import           Control.Monad (guard)
 
-import           Data.Bifunctor
-import           Data.Bifoldable
-import           Data.Bitraversable
-import           Data.Foldable (Foldable(foldr))
-#if !(MIN_VERSION_base(4,8,0))
-import           Data.Foldable (foldMap)
-#endif
-import           Data.Function (on)
-import           Data.List hiding (foldr)
-import qualified Data.Map as Map (fromList, lookup)
-import           Data.Map (Map)
+import           Data.Bifunctor.TH.Internal
+import           Data.List
 import           Data.Maybe
-import           Data.Monoid
+#if __GLASGOW_HASKELL__ < 710 && MIN_VERSION_template_haskell(2,8,0)
 import qualified Data.Set as Set
-import           Data.Set (Set)
-#if !(MIN_VERSION_base(4,8,0))
-import           Data.Traversable (Traversable(traverse))
 #endif
 
 import           Language.Haskell.TH.Lib
 import           Language.Haskell.TH.Ppr
 import           Language.Haskell.TH.Syntax
-
-import           Prelude hiding (foldr)
 
 -------------------------------------------------------------------------------
 -- User-facing API
@@ -201,7 +185,10 @@ deriveBifoldable = deriveBiClass Bifoldable
 -- | Generates a lambda expression which behaves like 'bifold' (without requiring a
 -- 'Bifoldable' instance).
 makeBifold :: Name -> Q Exp
-makeBifold name = [| $(makeBifoldMap name) id id |]
+makeBifold name = appsE [ makeBifoldMap name
+                        , varE idValName
+                        , varE idValName
+                        ]
 
 -- | Generates a lambda expression which behaves like 'bifoldMap' (without requiring a
 -- 'Bifoldable' instance).
@@ -216,10 +203,26 @@ makeBifoldr = makeBiFun Bifoldr
 -- | Generates a lambda expression which behaves like 'bifoldl' (without requiring a
 -- 'Bifoldable' instance).
 makeBifoldl :: Name -> Q Exp
-makeBifoldl name =
-    [| \f g z t -> appEndo (getDual ($(makeBifoldMap name) (Dual . Endo . flip f)
-                                                           (Dual . Endo . flip g)
-                                                           t)) z |]
+makeBifoldl name = do
+    f <- newName "f"
+    g <- newName "g"
+    z <- newName "z"
+    t <- newName "t"
+    lamE [varP f, varP g, varP z, varP t] $
+         appsE [ varE appEndoValName
+               , appsE [ varE getDualValName
+                       , appsE [ makeBifoldMap name, foldFun f, foldFun g, varE t]
+                       ]
+               , varE z
+               ]
+  where
+    foldFun :: Name -> Q Exp
+    foldFun n = infixApp (conE dualDataName)
+                         (varE composeValName)
+                         (infixApp (conE endoDataName)
+                                   (varE composeValName)
+                                   (varE flipValName `appE` varE n)
+                         )
 
 -- | Generates a 'Bitraversable' instance declaration for the given data type or data
 -- family instance.
@@ -234,18 +237,30 @@ makeBitraverse = makeBiFun Bitraverse
 -- | Generates a lambda expression which behaves like 'bisequenceA' (without requiring a
 -- 'Bitraversable' instance).
 makeBisequenceA :: Name -> Q Exp
-makeBisequenceA name = [| $(makeBitraverse name) id id |]
+makeBisequenceA name = appsE [ makeBitraverse name
+                             , varE idValName
+                             , varE idValName
+                             ]
 
 -- | Generates a lambda expression which behaves like 'bimapM' (without requiring a
 -- 'Bitraversable' instance).
 makeBimapM :: Name -> Q Exp
-makeBimapM name =
-    [| \f g -> unwrapMonad . $(makeBitraverse name) (WrapMonad . f) (WrapMonad . g) |]
+makeBimapM name = do
+    f <- newName "f"
+    g <- newName "g"
+    lamE [varP f, varP g] . infixApp (varE unwrapMonadValName) (varE composeValName) $
+                            appsE [makeBitraverse name, wrapMonadExp f, wrapMonadExp g]
+  where
+    wrapMonadExp :: Name -> Q Exp
+    wrapMonadExp n = infixApp (conE wrapMonadDataName) (varE composeValName) (varE n)
 
 -- | Generates a lambda expression which behaves like 'bisequence' (without requiring a
 -- 'Bitraversable' instance).
 makeBisequence :: Name -> Q Exp
-makeBisequence name = [| $(makeBimapM name) id id |]
+makeBisequence name = appsE [ makeBimapM name
+                            , varE idValName
+                            , varE idValName
+                            ]
 
 -------------------------------------------------------------------------------
 -- Code generation
@@ -367,7 +382,7 @@ makeBiFunForCons biFun nbs cons = do
         . appsE
         $ [ varE $ biFunConstName biFun
           , if null cons
-               then appE (varE 'error)
+               then appE (varE errorValName)
                          (stringE $ "Void " ++ nameBase (biFunName biFun))
                else caseE (varE value)
                           (map (makeBiFunForCon biFun z tvis) cons)
@@ -473,10 +488,18 @@ makeBiFunForType biFun tvis conName covariant ty =
              ArrowT | not (allowFunTys (biFunToClass biFun)) -> noFunctionsError conName
                     | mentionsTyArgs ->
                  let [argTy, resTy] = tyArgs
-                  in [| \x b ->
-                         $(makeBiFunForType biFun tvis conName covariant resTy)
-                         (x ($(makeBiFunForType biFun tvis conName (not covariant) argTy) b))
-                      |]
+                  in do x <- newName "x"
+                        b <- newName "b"
+                        lamE [varP x, varP b] $
+                             covBiFun covariant resTy `appE` (varE x `appE`
+                                (covBiFun (not covariant) argTy `appE` varE b))
+                          where
+                            covBiFun :: Bool -> Type -> Q Exp
+                            covBiFun cov t = makeBiFunForType biFun tvis conName cov t
+--                       [| \x b ->
+--                          $(makeBiFunForType biFun tvis conName covariant resTy)
+--                          (x ($(makeBiFunForType biFun tvis conName (not covariant) argTy) b))
+--                       |]
              TupleT n | n > 0 && mentionsTyArgs -> do
                  args <- mapM newName $ catMaybes [ Just "x"
                                                   , guard (biFun == Bifoldr) >> Just "z"
@@ -812,53 +835,6 @@ dataConIError = error
 #endif
 
 -------------------------------------------------------------------------------
--- Expanding type synonyms
--------------------------------------------------------------------------------
-
--- | Expands all type synonyms in a type. Written by Dan Rosén in the
--- @genifunctors@ package (licensed under BSD3).
-expandSyn :: Type -> Q Type
-expandSyn (ForallT tvs ctx t) = fmap (ForallT tvs ctx) $ expandSyn t
-expandSyn t@AppT{}            = expandSynApp t []
-expandSyn t@ConT{}            = expandSynApp t []
-expandSyn (SigT t _)          = expandSyn t   -- Ignore kind synonyms
-expandSyn t                   = return t
-
-expandSynApp :: Type -> [Type] -> Q Type
-expandSynApp (AppT t1 t2) ts = do
-    t2' <- expandSyn t2
-    expandSynApp t1 (t2':ts)
-expandSynApp (ConT n) ts | nameBase n == "[]" = return $ foldl' AppT ListT ts
-expandSynApp t@(ConT n) ts = do
-    info <- reify n
-    case info of
-        TyConI (TySynD _ tvs rhs) ->
-            let (ts', ts'') = splitAt (length tvs) ts
-                subs = mkSubst tvs ts'
-                rhs' = subst subs rhs
-             in expandSynApp rhs' ts''
-        _ -> return $ foldl' AppT t ts
-expandSynApp t ts = do
-    t' <- expandSyn t
-    return $ foldl' AppT t' ts
-
-type Subst = Map Name Type
-
-mkSubst :: [TyVarBndr] -> [Type] -> Subst
-mkSubst vs ts =
-   let vs' = map un vs
-       un (PlainTV v)    = v
-       un (KindedTV v _) = v
-   in Map.fromList $ zip vs' ts
-
-subst :: Subst -> Type -> Type
-subst subs (ForallT v c t) = ForallT v c $ subst subs t
-subst subs t@(VarT n)      = fromMaybe t $ Map.lookup n subs
-subst subs (AppT t1 t2)    = AppT (subst subs t1) (subst subs t2)
-subst subs (SigT t k)      = SigT (subst subs t) k
-subst _ t                  = t
-
--------------------------------------------------------------------------------
 -- Class-specific constants
 -------------------------------------------------------------------------------
 
@@ -870,21 +846,21 @@ data BiFun = Bimap | Bifoldr | BifoldMap | Bitraverse
   deriving Eq
 
 biFunConstName :: BiFun -> Name
-biFunConstName Bimap      = 'bimapConst
-biFunConstName Bifoldr    = 'bifoldrConst
-biFunConstName BifoldMap  = 'bifoldMapConst
-biFunConstName Bitraverse = 'bitraverseConst
+biFunConstName Bimap      = bimapConstValName
+biFunConstName Bifoldr    = bifoldrConstValName
+biFunConstName BifoldMap  = bifoldMapConstValName
+biFunConstName Bitraverse = bitraverseConstValName
 
 biClassName :: BiClass -> Name
-biClassName Bifunctor     = ''Bifunctor
-biClassName Bifoldable    = ''Bifoldable
-biClassName Bitraversable = ''Bitraversable
+biClassName Bifunctor     = bifunctorTypeName
+biClassName Bifoldable    = bifoldableTypeName
+biClassName Bitraversable = bitraversableTypeName
 
 biFunName :: BiFun -> Name
-biFunName Bimap      = 'bimap
-biFunName Bifoldr    = 'bifoldr
-biFunName BifoldMap  = 'bifoldMap
-biFunName Bitraverse = 'bitraverse
+biFunName Bimap      = bimapValName
+biFunName Bifoldr    = bifoldrValName
+biFunName BifoldMap  = bifoldMapValName
+biFunName Bitraverse = bitraverseValName
 
 biClassToFuns :: BiClass -> [BiFun]
 biClassToFuns Bifunctor     = [Bimap]
@@ -898,17 +874,17 @@ biFunToClass BifoldMap  = Bifoldable
 biFunToClass Bitraverse = Bitraversable
 
 biClassConstraint :: BiClass -> Int -> Maybe Name
-biClassConstraint Bifunctor     1 = Just ''Functor
-biClassConstraint Bifoldable    1 = Just ''Foldable
-biClassConstraint Bitraversable 1 = Just ''Traversable
+biClassConstraint Bifunctor     1 = Just functorTypeName
+biClassConstraint Bifoldable    1 = Just foldableTypeName
+biClassConstraint Bitraversable 1 = Just traversableTypeName
 biClassConstraint biClass       2 = Just $ biClassName biClass
 biClassConstraint _             _ = Nothing
 
 biFunArity :: BiFun -> Int -> Maybe Name
-biFunArity Bimap      1 = Just 'fmap
-biFunArity Bifoldr    1 = Just 'foldr
-biFunArity BifoldMap  1 = Just 'foldMap
-biFunArity Bitraverse 1 = Just 'traverse
+biFunArity Bimap      1 = Just fmapValName
+biFunArity Bifoldr    1 = Just foldrValName
+biFunArity BifoldMap  1 = Just foldMapValName
+biFunArity Bitraverse 1 = Just traverseValName
 biFunArity biFun      2 = Just $ biFunName biFun
 biFunArity _          _ = Nothing
 
@@ -922,14 +898,21 @@ allowExQuant _          = False
 
 -- See Trac #7436 for why explicit lambdas are used
 biFunTriv :: BiFun -> Q Exp
-biFunTriv Bimap      = [| \x   -> x      |]
-biFunTriv Bifoldr    = [| \_ z -> z      |]
-biFunTriv BifoldMap  = [| \_   -> mempty |]
-biFunTriv Bitraverse = [|         pure   |]
+biFunTriv Bimap = do
+    x <- newName "x"
+    lamE [varP x] $ varE x
+biFunTriv Bifoldr = do
+    z <- newName "z"
+    lamE [wildP, varP z] $ varE z
+biFunTriv BifoldMap = lamE [wildP] $ varE memptyValName
+biFunTriv Bitraverse = varE pureValName
 
 biFunApp :: BiFun -> Q Exp -> Q Exp
-biFunApp Bifoldr e = [| \x z -> $(e) z x |]
-biFunApp _       e = e
+biFunApp Bifoldr e = do
+    x <- newName "x"
+    z <- newName "z"
+    lamE [varP x, varP z] $ appsE [e, varE z, varE x]
+biFunApp _ e = e
 
 biFunCombine :: BiFun -> Name -> Name -> [Q Exp] -> Q Exp
 biFunCombine Bimap      = bimapCombine
@@ -944,282 +927,11 @@ bifoldrCombine :: Name -> Name -> [Q Exp] -> Q Exp
 bifoldrCombine _ zName = foldr appE (varE zName)
 
 bifoldMapCombine :: Name -> Name -> [Q Exp] -> Q Exp
-bifoldMapCombine _ _ [] = varE 'mempty
-bifoldMapCombine _ _ es = foldr1 (appE . appE (varE 'mappend)) es
+bifoldMapCombine _ _ [] = varE memptyValName
+bifoldMapCombine _ _ es = foldr1 (appE . appE (varE mappendValName)) es
 
 bitraverseCombine :: Name -> Name -> [Q Exp] -> Q Exp
-bitraverseCombine conName _ [] = varE 'pure `appE` conE conName
+bitraverseCombine conName _ [] = varE pureValName `appE` conE conName
 bitraverseCombine conName _ (e:es) =
-    foldl' (flip infixApp $ varE '(<*>)) (appsE [varE 'fmap, conE conName, e]) es
-
-bimapConst :: p b d -> (a -> b) -> (c -> d) -> p a c -> p b d
-bimapConst = const . const . const
-{-# INLINE bimapConst #-}
-
-bifoldrConst :: c -> (a -> c -> c) -> (b -> c -> c) -> c -> p a b -> c
-bifoldrConst = const . const . const . const
-{-# INLINE bifoldrConst #-}
-
-bifoldMapConst :: m -> (a -> m) -> (b -> m) -> p a b -> m
-bifoldMapConst = const . const . const
-{-# INLINE bifoldMapConst #-}
-
-bitraverseConst :: f (t c d) -> (a -> f c) -> (b -> f d) -> t a b -> f (t c d)
-bitraverseConst = const . const . const
-{-# INLINE bitraverseConst #-}
-
--------------------------------------------------------------------------------
--- NameBase
--------------------------------------------------------------------------------
-
--- | A wrapper around Name which only uses the 'nameBase' (not the entire Name)
--- to compare for equality. For example, if you had two Names a_123 and a_456,
--- they are not equal as Names, but they are equal as NameBases.
---
--- This is useful when inspecting type variables, since a type variable in an
--- instance context may have a distinct Name from a type variable within an
--- actual constructor declaration, but we'd want to treat them as the same
--- if they have the same 'nameBase' (since that's what the programmer uses to
--- begin with).
-newtype NameBase = NameBase { getName :: Name }
-
-getNameBase :: NameBase -> String
-getNameBase = nameBase . getName
-
-instance Eq NameBase where
-    (==) = (==) `on` getNameBase
-
-instance Ord NameBase where
-    compare = compare `on` getNameBase
-
-instance Show NameBase where
-    showsPrec p = showsPrec p . getNameBase
-
--- | A NameBase paired with the name of its map function. For example, when deriving
--- Bifunctor, its list of TyVarInfos might look like [(a, 'f), (b, 'g)].
-type TyVarInfo = (NameBase, Name)
-
--------------------------------------------------------------------------------
--- Assorted utilities
--------------------------------------------------------------------------------
-
-thd3 :: (a, b, c) -> c
-thd3 (_, _, c) = c
-
--- | Extracts the name of a constructor.
-constructorName :: Con -> Name
-constructorName (NormalC name      _  ) = name
-constructorName (RecC    name      _  ) = name
-constructorName (InfixC  _    name _  ) = name
-constructorName (ForallC _    _    con) = constructorName con
-
--- | Generate a list of fresh names with a common prefix, and numbered suffixes.
-newNameList :: String -> Int -> Q [Name]
-newNameList prefix n = mapM (newName . (prefix ++) . show) [1..n]
-
--- | Remove any occurrences of a forall-ed type variable from a list of @TyVarInfo@s.
-removeForalled :: [TyVarBndr] -> [TyVarInfo] -> [TyVarInfo]
-removeForalled tvbs = filter (not . foralled tvbs)
-  where
-    foralled :: [TyVarBndr] -> TyVarInfo -> Bool
-    foralled tvbs' tvi = fst tvi `elem` map (NameBase . tvbName) tvbs'
-
--- | Extracts the name from a TyVarBndr.
-tvbName :: TyVarBndr -> Name
-tvbName (PlainTV  name)   = name
-tvbName (KindedTV name _) = name
-
--- | Extracts the kind from a TyVarBndr.
-tvbKind :: TyVarBndr -> Kind
-tvbKind (PlainTV  _)   = starK
-tvbKind (KindedTV _ k) = k
-
--- | Replace the Name of a TyVarBndr with one from a Type (if the Type has a Name).
-replaceTyVarName :: TyVarBndr -> Type -> TyVarBndr
-replaceTyVarName tvb            (SigT t _) = replaceTyVarName tvb t
-replaceTyVarName (PlainTV  _)   (VarT n)   = PlainTV  n
-replaceTyVarName (KindedTV _ k) (VarT n)   = KindedTV n k
-replaceTyVarName tvb            _          = tvb
-
--- | Applies a typeclass constraint to a type.
-applyClass :: Name -> Name -> Pred
-#if MIN_VERSION_template_haskell(2,10,0)
-applyClass con t = AppT (ConT con) (VarT t)
-#else
-applyClass con t = ClassP con [VarT t]
-#endif
-
--- | Checks to see if the last types in a data family instance can be safely eta-
--- reduced (i.e., dropped), given the other types. This checks for three conditions:
---
--- (1) All of the dropped types are type variables
--- (2) All of the dropped types are distinct
--- (3) None of the remaining types mention any of the dropped types
-canEtaReduce :: [Type] -> [Type] -> Bool
-canEtaReduce remaining dropped =
-       all isTyVar dropped
-    && allDistinct nbs -- Make sure not to pass something of type [Type], since Type
-                       -- didn't have an Ord instance until template-haskell-2.10.0.0
-    && not (any (`mentionsNameBase` nbs) remaining)
-  where
-    nbs :: [NameBase]
-    nbs = map varTToNameBase dropped
-
--- | Extract the Name from a type variable.
-varTToName :: Type -> Name
-varTToName (VarT n)   = n
-varTToName (SigT t _) = varTToName t
-varTToName _          = error "Not a type variable!"
-
--- | Extract the NameBase from a type variable.
-varTToNameBase :: Type -> NameBase
-varTToNameBase = NameBase . varTToName
-
--- | Peel off a kind signature from a Type (if it has one).
-unSigT :: Type -> Type
-unSigT (SigT t _) = t
-unSigT t          = t
-
--- | Is the given type a variable?
-isTyVar :: Type -> Bool
-isTyVar (VarT _)   = True
-isTyVar (SigT t _) = isTyVar t
-isTyVar _          = False
-
--- | Is the given type a type family constructor (and not a data family constructor)?
-isTyFamily :: Type -> Q Bool
-isTyFamily (ConT n) = do
-    info <- reify n
-    return $ case info of
-#if MIN_VERSION_template_haskell(2,7,0)
-         FamilyI (FamilyD TypeFam _ _ _) _ -> True
-#else
-         TyConI  (FamilyD TypeFam _ _ _)   -> True
-#endif
-         _ -> False
-isTyFamily _ = return False
-
--- | Are all of the items in a list (which have an ordering) distinct?
---
--- This uses Set (as opposed to nub) for better asymptotic time complexity.
-allDistinct :: Ord a => [a] -> Bool
-allDistinct = allDistinct' Set.empty
-  where
-    allDistinct' :: Ord a => Set a -> [a] -> Bool
-    allDistinct' uniqs (x:xs)
-        | x `Set.member` uniqs = False
-        | otherwise            = allDistinct' (Set.insert x uniqs) xs
-    allDistinct' _ _           = True
-
--- | Does the given type mention any of the NameBases in the list?
-mentionsNameBase :: Type -> [NameBase] -> Bool
-mentionsNameBase = go Set.empty
-  where
-    go :: Set NameBase -> Type -> [NameBase] -> Bool
-    go foralls (ForallT tvbs _ t) nbs =
-        go (foralls `Set.union` Set.fromList (map (NameBase . tvbName) tvbs)) t nbs
-    go foralls (AppT t1 t2) nbs = go foralls t1 nbs || go foralls t2 nbs
-    go foralls (SigT t _)   nbs = go foralls t nbs
-    go foralls (VarT n)     nbs = varNb `elem` nbs && not (varNb `Set.member` foralls)
-      where
-        varNb = NameBase n
-    go _       _            _   = False
-
--- | Does an instance predicate mention any of the NameBases in the list?
-predMentionsNameBase :: Pred -> [NameBase] -> Bool
-#if MIN_VERSION_template_haskell(2,10,0)
-predMentionsNameBase = mentionsNameBase
-#else
-predMentionsNameBase (ClassP _ tys) nbs = any (`mentionsNameBase` nbs) tys
-predMentionsNameBase (EqualP t1 t2) nbs = mentionsNameBase t1 nbs || mentionsNameBase t2 nbs
-#endif
-
--- | The number of arrows that compose the spine of a kind signature
--- (e.g., (* -> *) -> k -> * has two arrows on its spine).
-numKindArrows :: Kind -> Int
-numKindArrows k = length (uncurryKind k) - 1
-
--- | Construct a type via curried application.
-applyTy :: Type -> [Type] -> Type
-applyTy = foldl' AppT
-
--- | Fully applies a type constructor to its type variables.
-applyTyCon :: Name -> [Type] -> Type
-applyTyCon = applyTy . ConT
-
--- | Split an applied type into its individual components. For example, this:
---
--- @
--- Either Int Char
--- @
---
--- would split to this:
---
--- @
--- [Either, Int, Char]
--- @
-unapplyTy :: Type -> [Type]
-unapplyTy = reverse . go
-  where
-    go :: Type -> [Type]
-    go (AppT t1 t2) = t2:go t1
-    go (SigT t _)   = go t
-    go t            = [t]
-
--- | Split a type signature by the arrows on its spine. For example, this:
---
--- @
--- (Int -> String) -> Char -> ()
--- @
---
--- would split to this:
---
--- @
--- [Int -> String, Char, ()]
--- @
-uncurryTy :: Type -> [Type]
-uncurryTy (AppT (AppT ArrowT t1) t2) = t1:uncurryTy t2
-uncurryTy (SigT t _)                 = uncurryTy t
-uncurryTy t                          = [t]
-
--- | Like uncurryType, except on a kind level.
-uncurryKind :: Kind -> [Kind]
-#if MIN_VERSION_template_haskell(2,8,0)
-uncurryKind = uncurryTy
-#else
-uncurryKind (ArrowK k1 k2) = k1:uncurryKind k2
-uncurryKind k              = [k]
-#endif
-
-wellKinded :: [Kind] -> Bool
-wellKinded = all canRealizeKindStar
-
--- | Of form k1 -> k2 -> ... -> kn, where k is either a single kind variable or *.
-canRealizeKindStarChain :: Kind -> Bool
-canRealizeKindStarChain = all canRealizeKindStar . uncurryKind
-
-canRealizeKindStar :: Kind -> Bool
-canRealizeKindStar k = case uncurryKind k of
-    [k'] -> case k' of
-#if MIN_VERSION_template_haskell(2,8,0)
-                 StarT    -> True
-                 (VarT _) -> True -- Kind k can be instantiated with *
-#else
-                 StarK    -> True
-#endif
-                 _ -> False
-    _ -> False
-
-# if MIN_VERSION_template_haskell(2,8,0) && __GLASGOW_HASKELL__ < 710
-distinctKindVars :: Kind -> Set Name
-distinctKindVars (AppT k1 k2) = distinctKindVars k1 `Set.union` distinctKindVars k2
-distinctKindVars (SigT k _)   = distinctKindVars k
-distinctKindVars (VarT k)     = Set.singleton k
-distinctKindVars _            = Set.empty
-#endif
-
-#if __GLASGOW_HASKELL__ >= 708 && __GLASGOW_HASKELL__ < 710
-tvbToType :: TyVarBndr -> Type
-tvbToType (PlainTV n)    = VarT n
-tvbToType (KindedTV n k) = SigT (VarT n) k
-#endif
+    foldl' (flip infixApp $ varE apValName)
+        (appsE [varE fmapValName, conE conName, e]) es
